Avoid breaking drift
